Change root vector value and set it to 1
This commit modifies the code in `Tree.hs` by adding a new function `changeRootVectorValue` that changes the value of the root vector in a component. The function `setRootVectorTo1` is then defined as a specific case of `changeRootVectorValue`, setting the root vector value to 1. The changes also include adding sample calls to these functions and printing the result, as well as an updated example tree.